+++ src/algoritmia/algorithms/topological_sort.py
@@ -2,9 +2,14 @@ from typing import Optional
 from collections.abc import Iterator
 from algoritmia.datastructures.graphs import Digraph, TVertex
 
+# Ordenación topológica de un digrafo acíclico
+# Si el grafo tiene algún ciclo, el algoritmo no funcionará bien, puede:
+#  - Lanzar la excepción 'RecursionError: maximum recursion depth exceeded'
+#    (si el ciclo afecta a traverse_from())
+#  - Devolver None (si el ciclo afecta a next_vertex_no_preds())
+
 def topological_sort(g: Digraph[TVertex]) -> Optional[list[TVertex]]:
     def traverse_from(v) -> Iterator[TVertex]:  # depthfirst postorder
-        pending.remove(v)
         for suc in g.succs(v):
             if suc in pending:
                 yield from traverse_from(suc)
@@ -20,8 +25,8 @@ def topological_sort(g: Digraph[TVertex]) -> Optional[list[TVertex]]:
     while len(pending) > 0:
         u = next_vertex_no_preds()
         if u is None:
-            return None
-        vertices = list(traverse_from(u))
+            return None  # No hay ningún vertice sin predecesores: ciclo
+        vertices = list(traverse_from(u))  # Un ciclo provocaría recursividad infinita en traverse_from()
         pending -= set(vertices)
         lv.extend(vertices)
     lv.reverse()