+++ src/algoritmia/examples/greedy/activities_selector_greedy.py
@@ -3,11 +3,11 @@ Interval = tuple[int, int]  # (start, end)  both are >=0
 # Devuelve una lista con el máximo de actividades posibles sin solapes
 def select(activities: list[Interval]) -> list[Interval]:
     # Ordenamos los índices por hora de finalización de la tarea
-    sorted_index = sorted(range(len(activities)), key=lambda i: activities[i][1])
+    sorted_indices = sorted(range(len(activities)), key=lambda i: activities[i][1])
 
     res: list[Interval] = []
     last_task_end = 0
-    for i in sorted_index:
+    for i in sorted_indices:
         start, end = activities[i]
         if last_task_end <= start:
             res.append(activities[i])