+++ src/algoritmia/examples/greedy/coin_change_greedy.py
@@ -2,9 +2,9 @@ from typing import Optional
 
 
 def coin_change_solve_naif(v: tuple[int, ...], Q: int) -> Optional[list[int]]:
-    res: list[int] = []
+    res: list[int] = [0] * len(v)
     for i in range(len(v)):
-        res.append(Q // v[i])
+        res[i] = Q // v[i]
         Q = Q % v[i]
     if Q == 0:
         return res
@@ -27,18 +27,14 @@ def coin_change_solve(v: tuple[int, ...], Q: int) -> Optional[list[int]]:
 
 
 if __name__ == '__main__':
-    print(coin_change_solve((1, 2, 5, 10), 6))  # Ok: [1, 0, 1, 0]
-    print(coin_change_solve((1, 9, 15), 19))  # Mal: [4, 0, 1]
-    print(coin_change_solve((2, 9, 15), 10))  # Mal: None
-
-
-if __name__ == '__main__':
-    print(coin_change_solve_naif((1, 2, 5, 10), 6))  # [6, 0, 0, 0]
-    print(coin_change_solve_naif((2, 5, 10), 7))  # None
-    print(coin_change_solve_naif((1, 9, 15), 19))  # [19, 0, 0]
-    print(coin_change_solve_naif((2, 9, 15), 10))  # [5, 0, 0]
+    print("coin_change_solve_naif:")
+    print(coin_change_solve_naif((1, 2, 5, 10), 6))  # Mal: [6, 0, 0, 0]
+    print(coin_change_solve_naif((2, 5, 10), 7))     # Mal: None
+    print(coin_change_solve_naif((1, 9, 15), 19))    # Mal: [19, 0, 0]
+    print(coin_change_solve_naif((2, 9, 15), 10))    # Mal: [5, 0, 0]
     print()
-    print(coin_change_solve((1, 2, 5, 10), 6))  # [1, 0, 1, 0]
-    print(coin_change_solve((2, 5, 10), 7))  # [1, 1, 0]]
-    print(coin_change_solve((1, 9, 15), 19))  # [4, 0, 1]
-    print(coin_change_solve((2, 9, 15), 10))  # None
+    print("coin_change_solve:")
+    print(coin_change_solve((1, 2, 5, 10), 6))  # Ok: [1, 0, 1, 0]
+    print(coin_change_solve((2, 5, 10), 7))     # Ok: [1, 1, 0]]
+    print(coin_change_solve((1, 9, 15), 19))    # Mal: [4, 0, 1]
+    print(coin_change_solve((2, 9, 15), 10))    # Mal: None