+++ tests/datastructures/test_graphs.py
@@ -13,6 +13,7 @@ class TestDigraphs(unittest.TestCase):
         self.g1bis = Digraph(E=self.edges)
         self.g1bis.add_vertex(6)
         edges = self.edges + [(1, 3)]
+        self.g6 = UndirectedGraph(V=range(8), E=self.edges)
         self.g8 = Digraph(V=range(8), E=edges)
 
     def test_readonly(self):
@@ -37,14 +38,56 @@ class TestDigraphs(unittest.TestCase):
 
     def test_edit_edges(self):
         for g in self.g1, self.g1bis:
+            self.assertFalse(1 in g.succs(7))
+            self.assertFalse(7 in g.succs(1))
+            self.assertFalse(7 in g.preds(1))
+            self.assertFalse(1 in g.preds(7))
             g.add_edge((7, 1))
             self.assertEqual(len(g.E), len(self.edges) + 1)
             self.assertTrue((7, 1) in g.E)
             self.assertFalse((1, 7) in g.E)
+            self.assertTrue(1 in g.succs(7))
+            self.assertFalse(7 in g.succs(1))
+            self.assertTrue(7 in g.preds(1))
+            self.assertFalse(1 in g.preds(7))
+
+            self.assertTrue(7 in g.succs(0))
+            self.assertFalse(0 in g.succs(7))
+            self.assertTrue(0 in g.preds(7))
+            self.assertFalse(7 in g.preds(0))
             g.remove_edge((0, 7))
             self.assertFalse((0, 7) in g.E)
             self.assertEqual(len(g.E), len(self.edges))
+            self.assertFalse(0 in g.succs(7))
+            self.assertFalse(7 in g.succs(0))
+            self.assertFalse(7 in g.preds(0))
+            self.assertFalse(0 in g.preds(7))
 
+    def test_edit_vertices(self):
+        for g in self.g1, self.g1bis:
+            g.add_vertex(6)
+            self.assertTrue(6 in g.V)
+            self.assertFalse((0, 6) in g.E)
+            self.assertFalse(0 in g.succs(6))
+            self.assertFalse((6, 0) in g.E)
+            self.assertFalse(6 in g.succs(0))
+
+            if g == self.g6:  # Por cobertura
+                g.add_vertex(10)
+
+            g.remove_vertex(0)
+            self.assertFalse((1, 0) in g.E)
+            self.assertFalse((0, 1) in g.E)
+            self.assertFalse(0 in set(s for v in g.V for s in g.succs(v)))
+            self.assertFalse(0 in set(s for v in g.V for s in g.preds(v)))
+
+    def test_contains(self):
+        for g in self.g1, self.g1bis:
+            self.assertFalse(g.contains_edge((7, 1)))
+            self.assertFalse(g.contains_edge((1, 7)))
+            g.add_edge((7, 1))
+            self.assertTrue(g.contains_edge((7, 1)))
+            self.assertFalse(g.contains_edge((1, 7)))
 
 class TestUndirectedGraph(unittest.TestCase):
     def setUp(self):
@@ -118,21 +161,36 @@ class TestUndirectedGraph(unittest.TestCase):
             # self.assertEqual(len(g.E), len(self.undirected_edges)+2)
             self.assertTrue((7, 1) in g.E)
             # self.assertTrue((1,7) in g.E)
+            self.assertTrue(7 in g.succs(1))
+            self.assertTrue(1 in g.succs(7))
+            self.assertTrue(7 in g.preds(1))
+            self.assertTrue(1 in g.preds(7))
+
 
             g.remove_edge((0, 7))
             # self.assertFalse((7,0) in g.E)
             self.assertFalse((0, 7) in g.E)
+            self.assertFalse(7 in g.succs(0))
+            self.assertFalse(0 in g.succs(7))
+            self.assertFalse(7 in g.preds(0))
+            self.assertFalse(0 in g.preds(7))
 
             g.add_edge((1, 3))
             self.assertTrue((1, 3) in g.E)
             # self.assertTrue((3,1) in g.E)
+            self.assertTrue(1 in g.succs(3))
+            self.assertTrue(3 in g.succs(1))
+            self.assertTrue(1 in g.preds(3))
+            self.assertTrue(3 in g.preds(1))
 
     def test_edit_vertices(self):
         for g in self.g1, self.g1bis:
             g.add_vertex(6)
             self.assertTrue(6 in g.V)
             self.assertFalse((0, 6) in g.E)
+            self.assertFalse(0 in g.succs(6))
             self.assertFalse((6, 0) in g.E)
+            self.assertFalse(6 in g.succs(0))
 
             if g == self.g6:  # Por cobertura
                 g.add_vertex(10)
@@ -140,7 +198,16 @@ class TestUndirectedGraph(unittest.TestCase):
             g.remove_vertex(0)
             self.assertFalse((1, 0) in g.E)
             self.assertFalse((0, 1) in g.E)
+            self.assertFalse(0 in set(s for v in g.V for s in g.succs(v)))
+            self.assertFalse(0 in set(s for v in g.V for s in g.preds(v)))
 
+    def test_contains(self):
+        for g in self.g1, self.g1bis:
+            self.assertFalse(g.contains_edge((7, 1)))
+            self.assertFalse(g.contains_edge((1, 7)))
+            g.add_edge((7, 1))
+            self.assertTrue(g.contains_edge((7, 1)))
+            self.assertTrue(g.contains_edge((1, 7)))
 
 class TestRedimDigraphs(unittest.TestCase):
     def setUp(self):
@@ -169,17 +236,27 @@ class TestRedimDigraphs(unittest.TestCase):
 class TestWeightingFunction(unittest.TestCase):
     def setUp(self):
         self.sym = WeightingFunction([((0, 0), 1), ((1, 1), 2), ((1, 2), 3)], symmetrical=True)
+        self.sym_d = WeightingFunction({(0, 0): 1, (1, 1): 2, (1, 2): 3}, symmetrical=True)
         self.asym = WeightingFunction([((0, 0), 1), ((1, 1), 2), ((1, 2), 3)], symmetrical=False)
+        self.asym_d = WeightingFunction({(0, 0): 1, (1, 1): 2, (1, 2): 3}, symmetrical=False)
+        self.rep = WeightingFunction([((0, 1), 10), ((0, 2), 2), ((1, 0), 10)], symmetrical=True)
 
     def test_getitem(self):
-        self.assertEqual(self.sym(0, 0), 1)
-        self.assertEqual(self.sym(1, 1), 2)
-        self.assertEqual(self.sym(1, 2), 3)
-        self.assertEqual(self.sym(2, 1), 3)
-        self.assertEqual(self.asym(0, 0), 1)
-        self.assertEqual(self.asym(1, 1), 2)
-        self.assertEqual(self.asym(1, 2), 3)
-        self.assertRaises(KeyError, self.asym.__call__, (2, 1))
+        for wf in [self.sym, self.sym_d]:
+            self.assertEqual(wf(0, 0), 1)
+            self.assertEqual(wf(1, 1), 2)
+            self.assertEqual(wf(1, 2), 3)
+            self.assertEqual(wf(2, 1), 3)
+        for wf in [self.asym, self.asym_d]:
+            self.assertEqual(wf(0, 0), 1)
+            self.assertEqual(wf(1, 1), 2)
+            self.assertEqual(wf(1, 2), 3)
+            self.assertRaises(KeyError, wf.__call__, (2, 1))
+            self.assertRaises(KeyError, wf.__call__, (2, 1))
+
+        self.assertRaises(ValueError, WeightingFunction, *[[((0, 1), 1), ((0, 2), 2), ((1, 0), 3)], True])
+        self.assertEqual(self.rep(0, 1), 10)
+        self.assertEqual(self.rep(1, 0), 10)
 
 
 if __name__ == "__main__":